ai-backend/requirements.txt
fastapi
uvicorn
pydantic
pillow
pytesseract
langchain
openai
google-generativeai
scikit-learn
numpy
pandas
pymongo
python-multipart
python-dotenv
